move exception, fix tests and checkstyle
