+++ .gitignore
@@ -60,3 +60,4 @@ grakn-test/test-integration/benchmarks/
 
 # Python cache #
 __pycache__
+/local_scripts/
