--- graph.in
@@ -1,124 +1,0 @@
-digraph mygraph {
-  node [shape=box];
-"//server:server"
-"//server:server" -> "//kb/concept/manager:manager"
-"//server:server" -> "//concept/answer:answer"
-"//server:server" -> "//concept/util/attribute:attribute"
-"//server:server" -> "//concept/impl:impl"
-"//server:server" -> "//kb/concept/structure:structure"
-"//server:server" -> "//concept/structure:structure"
-"//server:server" -> "//kb/concept/api:api"
-"//server:server" -> "//common:common"
-"//server:server" -> "//server:version"
-"//server:server" -> "//graph-hadoop:graph-hadoop"
-"//server:server" -> "//graph-cql:graph-cql"
-"//server:server" -> "@graknlabs_benchmark//lib:server-instrumentation"
-"//server:server" -> "@graknlabs_common//:common"
-"//server:server" -> "//graph:graph"
-"//server:server" -> "//core:core"
-"//server:server" -> "//graql:graql"
-"//server:server" -> "//kb:kb"
-"//server:server" -> "//concept/manager:manager"
-"//server:server" -> "//concept/util:util"
-"//concept/manager:manager"
-"//concept/manager:manager" -> "//concept/util/attribute:attribute"
-"//concept/manager:manager" -> "//concept/structure:structure"
-"//concept/manager:manager" -> "//concept/impl:impl"
-"//concept/manager:manager" -> "//kb/concept/manager:manager"
-"//concept/manager:manager" -> "//core:core"
-"//concept/manager:manager" -> "//kb:kb"
-"//concept/manager:manager" -> "//kb/concept/api:api"
-"//concept/manager:manager" -> "//kb/concept/structure:structure"
-"//graph-hadoop:graph-hadoop"
-"//graph-hadoop:graph-hadoop" -> "//graph:graph"
-"//graph-hadoop:graph-hadoop" -> "//graph-cql:graph-cql"
-"//graph-hadoop:graph-hadoop" -> "//common:common"
-"//graql:graql"
-"//graql:graql" -> "//concept/util:util"
-"//graql:graql" -> "//core:core"
-"//graql:graql" -> "@graknlabs_benchmark//lib:server-instrumentation"
-"//graql:graql" -> "@graknlabs_common//:common"
-"//graql:graql" -> "//kb:kb"
-"//graql:graql" -> "//concept/util/attribute:attribute"
-"//graql:graql" -> "//concept/impl:impl"
-"//graql:graql" -> "//concept/answer:answer"
-"//graql:graql" -> "//kb/concept/manager:manager"
-"//graql:graql" -> "//kb/concept/api:api"
-"//graql:graql" -> "//common:common"
-"@graknlabs_benchmark//lib:server-instrumentation"
-"//graph-cql:graph-cql"
-"//graph-cql:graph-cql" -> "//graph:graph"
-"//kb:kb"
-"//kb:kb" -> "//kb/concept/structure:structure"
-"//kb:kb" -> "@graknlabs_common//:common"
-"//kb:kb" -> "//core:core"
-"//kb:kb" -> "//concept/answer:answer"
-"//kb:kb" -> "//kb/concept/manager:manager"
-"//kb:kb" -> "//kb/concept/api:api"
-"//kb:kb" -> "//common:common"
-"//concept/answer:answer"
-"//concept/answer:answer" -> "//kb/concept/api:api"
-"//concept/answer:answer" -> "//concept/util:util"
-"//concept/util:util"
-"//concept/util:util" -> "//kb/concept/structure:structure"
-"//concept/util:util" -> "//core:core"
-"//concept/util:util" -> "//concept/impl:impl"
-"//concept/util:util" -> "//kb/concept/api:api"
-"//concept/impl:impl"
-"//concept/impl:impl" -> "//concept/util/attribute:attribute"
-"//concept/impl:impl" -> "//concept/cache:conceptcache"
-"//concept/impl:impl" -> "//concept/structure:structure"
-"//concept/impl:impl" -> "//kb/concept/api:api"
-"//concept/impl:impl" -> "//kb/concept/structure:structure"
-"//concept/impl:impl" -> "//kb/concept/manager:manager"
-"//concept/impl:impl" -> "//core:core"
-"//concept/structure:structure"
-"//concept/structure:structure" -> "//graph:graph"
-"//concept/structure:structure" -> "//concept/cache:conceptcache"
-"//concept/structure:structure" -> "//core:core"
-"//concept/structure:structure" -> "//kb/concept/manager:manager"
-"//concept/structure:structure" -> "//kb/concept/structure:structure"
-"//concept/structure:structure" -> "//kb/concept/api:api"
-"//kb/concept/manager:manager"
-"//kb/concept/manager:manager" -> "//kb/concept/structure:structure"
-"//kb/concept/manager:manager" -> "//kb/concept/api:api"
-"//kb/concept/manager:manager" -> "//core:core"
-"//concept/cache:conceptcache"
-"//kb/concept/structure:structure"
-"//kb/concept/structure:structure" -> "//core:core"
-"//kb/concept/structure:structure" -> "//common:common"
-"//kb/concept/structure:structure" -> "//kb/concept/api:api"
-"//core:core"
-"//core:core" -> "@graknlabs_common//:common"
-"//core:core" -> "//graph:graph"
-"//core:core" -> "//kb/concept/api:api"
-"//core:core" -> "//common:common"
-"//graph:graph"
-"@info-picocli-picocli//jar:jar"
-"@info-picocli-picocli//jar:jar" -> "@info-picocli-picocli//jar:info-picocli-picocli.jar\n@info-picocli-picocli//jar:info-picocli-picocli-sources.jar"
-"@graknlabs_protocol//keyspace:keyspace-proto"
-"@graknlabs_protocol//keyspace:keyspace-proto" -> "@graknlabs_protocol//keyspace:Keyspace.proto"
-"@graknlabs_protocol//keyspace:Keyspace.proto"
-"@graknlabs_protocol//session:session-proto"
-"@graknlabs_protocol//session:session-proto" -> "@graknlabs_protocol//session:concept-proto"
-"@graknlabs_protocol//session:session-proto" -> "@graknlabs_protocol//session:answer-proto"
-"@graknlabs_protocol//session:session-proto" -> "@graknlabs_protocol//session:Session.proto"
-"@graknlabs_protocol//session:answer-proto"
-"@graknlabs_protocol//session:answer-proto" -> "@graknlabs_protocol//session:concept-proto"
-"@graknlabs_protocol//session:answer-proto" -> "@graknlabs_protocol//session:Answer.proto"
-"@graknlabs_protocol//session:Answer.proto"
-"@graknlabs_protocol//session:concept-proto"
-"@graknlabs_protocol//session:concept-proto" -> "@graknlabs_protocol//session:Concept.proto"
-"@graknlabs_protocol//session:Concept.proto"
-"@graknlabs_protocol//session:Session.proto"
-"//concept/util/attribute:attribute"
-"//concept/util/attribute:attribute" -> "@graknlabs_common//:common"
-"//concept/util/attribute:attribute" -> "//kb/concept/api:api"
-"//kb/concept/api:api"
-"//kb/concept/api:api" -> "@graknlabs_common//:common"
-"//kb/concept/api:api" -> "//common:common"
-"@graknlabs_graql//grammar:Graql.g4"
-"//common:common"
-"@graknlabs_common//:common"
-"//server:version"
-}
