+++ graql/executor/ConceptBuilderImpl.java
@@ -36,7 +36,7 @@ import grakn.core.kb.concept.manager.ConceptManager;
 import grakn.core.kb.graql.executor.ConceptBuilder;
 import grakn.core.kb.graql.executor.WriteExecutor;
 import grakn.core.kb.server.exception.GraqlQueryException;
-import grakn.core.kb.graql.GraqlSemanticException;
+import grakn.core.kb.graql.exception.GraqlSemanticException;
 import grakn.core.kb.server.exception.InvalidKBException;
 import graql.lang.Graql;
 import graql.lang.pattern.Pattern;
