+++ graql/executor/property/RelationExecutor.java
@@ -28,7 +28,7 @@ import grakn.core.kb.concept.api.Thing;
 import grakn.core.kb.graql.executor.WriteExecutor;
 import grakn.core.kb.graql.executor.property.PropertyExecutor;
 import grakn.core.kb.graql.gremlin.EquivalentFragmentSet;
-import grakn.core.kb.graql.GraqlSemanticException;
+import grakn.core.kb.graql.exception.GraqlSemanticException;
 import graql.lang.property.RelationProperty;
 import graql.lang.property.VarProperty;
 import graql.lang.statement.Statement;
