+++ graql/executor/property/ThenExecutor.java
@@ -23,7 +23,7 @@ import com.google.common.collect.ImmutableSet;
 import grakn.core.kb.graql.executor.WriteExecutor;
 import grakn.core.kb.graql.executor.property.PropertyExecutor;
 import grakn.core.kb.graql.gremlin.EquivalentFragmentSet;
-import grakn.core.kb.graql.GraqlSemanticException;
+import grakn.core.kb.graql.exception.GraqlSemanticException;
 import graql.lang.property.ThenProperty;
 import graql.lang.property.VarProperty;
 import graql.lang.statement.Variable;
