+++ kb/graql/exception/GraqlSemanticException.java
@@ -17,7 +17,7 @@
  *
  */
 
-package grakn.core.kb.graql;
+package grakn.core.kb.graql.exception;
 
 import grakn.core.common.exception.ErrorMessage;
 import grakn.core.common.exception.GraknException;
