+++ kb/graql/executor/WriteExecutor.java
@@ -21,7 +21,7 @@ package grakn.core.kb.graql.executor;
 
 import grakn.core.concept.answer.ConceptMap;
 import grakn.core.kb.concept.api.Concept;
-import grakn.core.kb.graql.GraqlSemanticException;
+import grakn.core.kb.graql.exception.GraqlSemanticException;
 import graql.lang.statement.Statement;
 import graql.lang.statement.Variable;
