+++ test-integration/graql/query/QueryBuilderIT.java
@@ -21,7 +21,7 @@ package grakn.core.graql.query;
 import grakn.core.graql.graph.MovieGraph;
 import grakn.core.kb.server.Session;
 import grakn.core.kb.server.Transaction;
-import grakn.core.kb.graql.GraqlSemanticException;
+import grakn.core.kb.graql.exception.GraqlSemanticException;
 import grakn.core.rule.GraknTestServer;
 import graql.lang.Graql;
 import graql.lang.query.GraqlDelete;
